% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_group_means.R
\name{estimate_group_means}
\alias{estimate_group_means}
\title{For each protein estimate latent intensity per condition}
\usage{
estimate_group_means(X, design, d0, s0, location_est, scale_est)
}
\arguments{
\item{X}{a numerical matrix that contains the log intensity values. Missing
data has to be encoded as 0}

\item{design}{a design matrix that has as many columns as conditions
and as many rows as samples (i.e. columns) in X. If a sample belongs
to a condition it has value 1, if not the matrix contains a 0.}

\item{d0}{estimated degrees of freedom of the variance moderation prior. Can be found
using \code{estimate_variance_moderation}}

\item{s0}{estimated variance of the variance moderation prior. Can be found
using \code{estimate_variance_moderation}}

\item{location_est}{inflection point of the sigmoid. Can be calculated using
\code{estimate_sigmoid}}

\item{scale_est}{steepness of the sigmoid. Usually negative and can be calculate
using \code{estimate_sigmoid}}
}
\description{
MLE for the group mean of each condition in each protein, taking into account
the likelihood of observing a zero (quantified by the sigmoid).
The function can also deal with all zero groups.
}
