% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sigmoid.R
\name{estimate_sigmoid}
\alias{estimate_sigmoid}
\title{Estimate dependency between missing values and intensity}
\usage{
estimate_sigmoid(X, data_description, d0, s0, n_iter = 2000, ...)
}
\arguments{
\item{X}{a numerical matrix that contains the log intensity values. Missing
data has to be encoded as 0}

\item{data_description}{a dataframe with a `Condition` factor column which 
assign each column of \code{X} to a condition}

\item{d0}{estimated degrees of freedom of the variance moderation prior. Can be found
using \code{estimate_variance_moderation}}

\item{s0}{estimated variance of the variance moderation prior. Can be found
using \code{estimate_variance_moderation}}

\item{n_iter}{the number of MCMC iterations}

\item{...}{additional parameters that are passed on to \code{rstan::sampling}}
}
\description{
Function that estimates the increasing likelihood to observe a missing
value if the latent protein intensity is lower. The decrease is parameterized
as a sigmoid and its location and scale are estimated. The function constructs
regularized mean estimates for each protein in each condition. Those estimates
are used to find the best sigmoid function that explains the drop outs (i.e. zeros).
Internally the function uses MCMC (implemented with Stan), it can thus take
a considerable time to calculate the results (1000x6 matrix took ~ 2h).
}
