% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_variance_moderation.R
\name{estimate_variance_moderation}
\alias{estimate_variance_moderation}
\title{Estimate variance moderation prior}
\usage{
estimate_variance_moderation(X, design)
}
\arguments{
\item{X}{a numerical matrix that contains the log intensity values. Missing
data can either be encoded as 0 or NA.}

\item{design}{a design matrix that has as many columns as conditions
and as many rows as samples (i.e. columns) in X. If a sample belongs
to a condition it has value 1, if not the matrix contains a 0.}
}
\description{
Function to estimate the parameters of an inverse chi-squared distribution
that is assummed as the prior of the variance for each protein.
The method was initially developed in Smyth, 2004 in the context of microarrays
to increase the power to detected differentially expressed genes. The
estimated inverse chi-squared distribution can be used to moderate the
variance estimate for each individual protein.
}
