% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliztions.R
\name{ma_plot}
\alias{ma_plot}
\title{MA Plot}
\usage{
ma_plot(X, design, reference_cond = 1)
}
\arguments{
\item{X}{a numerical matrix that contains the log intensity values. Missing
data has to be encoded as 0}

\item{design}{a design matrix that has as many columns as conditions
and as many rows as samples (i.e. columns) in X. If a sample belongs
to a condition it has value 1, if not the matrix contains a 0.}

\item{reference_cond}{the index of the design column with the condition
that is used as the reference.}
}
\description{
MA Plot
}
