% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pseudoLog10}
\alias{pseudoLog10}
\title{Function that will be transformed with a function
 that is very close to log10 for values between 1-infinity, but has 
 the feature that pseudoLog10(0) = 0 and pseudoLog10(- x) = - pseudoLog10(x)}
\usage{
pseudoLog10(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Function that will be transformed with a function
 that is very close to log10 for values between 1-infinity, but has 
 the feature that pseudoLog10(0) = 0 and pseudoLog10(- x) = - pseudoLog10(x)
}
